+++ tests/test_metagen.py
@@ -9,7 +9,10 @@ import pytest
 
 from pymetagen import MetaGen, json_metadata_to_pandas
 from pymetagen._typing import DataFrameT
-from pymetagen.datatypes import MetaGenSupportedLoadingMode
+from pymetagen.datatypes import (
+    MetaGenMetadataColumn,
+    MetaGenSupportedLoadingMode,
+)
 from pymetagen.exceptions import (
     FileTypeUnsupportedError,
     LoadingModeUnsupportedError,
@@ -47,11 +50,10 @@ class TestMetaGen:
 
     def test_compute_metadata(self, data: str, request: pytest.FixtureRequest):
         df: DataFrameT = request.getfixturevalue(data)
-        metadata = MetaGen(
-            data=df,
-        ).compute_metadata()
+        metagen = MetaGen(data=df)
+        metadata = metagen.compute_metadata()
 
-        assert set(metadata.index) == set(df.columns)
+        assert set(metadata.index) == set(metagen.columns)
 
     @pytest.mark.parametrize(
         "extension, read_metadata",
@@ -87,23 +89,10 @@ class TestMetaGen:
         if extension == "json":
             outdata.reset_index(names=["Name"], inplace=True)
 
-        assert len(outdata) == len(df.columns)
-        assert list(outdata.columns) == [
-            "Name",
-            "Long Name",
-            "Type",
-            "Description",
-            "Min",
-            "Max",
-            "Min Length",
-            "Max Length",
-            "# nulls",
-            "# empty/zero",
-            "# positive",
-            "# negative",
-            "# unique",
-            "Values",
-        ]
+        assert len(outdata) == metagen.columns_length
+        assert list(
+            outdata.columns
+        ) == MetaGenMetadataColumn.pymetagen_columns(include_name_column=True)
 
     @pytest.mark.parametrize(
         "column_name, expected_value",
@@ -230,7 +219,7 @@ class TestMetaGenFromPath:
         file_path: Path = request.getfixturevalue(path)
         MetaGen.from_path(
             path=file_path,
-            mode=mode,
+            loading_mode=mode,
         )
 
     def test_unsupported_path(
@@ -241,7 +230,7 @@ class TestMetaGenFromPath:
         with pytest.raises(FileTypeUnsupportedError):
             MetaGen.from_path(
                 path=tmp_dir_path / "test.unsupported",
-                mode=mode,
+                loading_mode=mode,
             )
 
     @pytest.mark.parametrize(
@@ -261,7 +250,9 @@ class TestMetaGenFromPath:
         description_path: Path = request.getfixturevalue(descriptions_path)
 
         metagen = MetaGen.from_path(
-            path=input_csv_path, descriptions_path=description_path, mode=mode
+            path=input_csv_path,
+            descriptions_path=description_path,
+            loading_mode=mode,
         )
         metadata = metagen.compute_metadata()
 
@@ -275,7 +266,7 @@ def test_from_path_unsupported_mode(tmp_dir_path: Path):
     with pytest.raises(LoadingModeUnsupportedError):
         MetaGen.from_path(
             path=tmp_dir_path / "test.csv",
-            mode="unsupported_mode",  # type: ignore
+            loading_mode="unsupported_mode",  # type: ignore
         )
 
 
@@ -291,7 +282,7 @@ def test_load_file_extension_none(
     with pytest.raises(FileTypeUnsupportedError):
         MetaGen.from_path(
             path=tmp_dir_path / "test",
-            mode=MetaGenSupportedLoadingMode.LAZY,
+            loading_mode=MetaGenSupportedLoadingMode.LAZY,
         )
 
 
@@ -299,7 +290,7 @@ def test_file_extension_none_for_parquet_directories():
     path = Path("tests/data/input_ab_partition")
     MetaGen.from_path(
         path=path,
-        mode=MetaGenSupportedLoadingMode.LAZY,
+        loading_mode=MetaGenSupportedLoadingMode.LAZY,
     )
 
 
@@ -308,7 +299,7 @@ def test_file_extension_none_for_directories_with_no_parquet_files():
     with pytest.raises(FileTypeUnsupportedError):
         MetaGen.from_path(
             path=path,
-            mode=MetaGenSupportedLoadingMode.LAZY,
+            loading_mode=MetaGenSupportedLoadingMode.LAZY,
         )
 
 
@@ -329,11 +320,11 @@ class TestMetaGenExtractData:
         inspection_mode: InspectionMode,
     ):
         metagen = MetaGen.from_path(
-            Path("tests/data/input_ab_partition.parquet"), mode=mode
+            Path("tests/data/input_ab_partition.parquet"), loading_mode=mode
         )
 
         extract = metagen.extract_data(
-            tbl_rows=2, mode=mode, inspection_mode=inspection_mode
+            tbl_rows=2, loading_mode=mode, inspection_mode=inspection_mode
         )
 
         assert extract.shape[0] == 2
@@ -358,12 +349,12 @@ class TestMetaGenWriteExtracts:
     ):
         metagen = MetaGen.from_path(
             path=Path("tests/data/input_ab_partition.parquet"),
-            mode=mode,
+            loading_mode=mode,
         )
         file_name = f"test-{inspection_mode}.csv"
         metagen.write_extract_by_inspection_mode(
             output_path=tmp_dir_path / file_name,
-            mode=mode,
+            loading_mode=mode,
             inspection_mode=inspection_mode,
             random_seed=None,
             number_rows=2,
@@ -383,12 +374,13 @@ class TestMetaGenWriteExtracts:
         mode: MetaGenSupportedLoadingMode,
     ):
         metagen = MetaGen.from_path(
-            path=Path("tests/data/input_ab_partition.parquet"), mode=mode
+            path=Path("tests/data/input_ab_partition.parquet"),
+            loading_mode=mode,
         )
         metagen.write_extracts(
             output_path=tmp_dir_path / "test.csv",
             number_rows=2,
-            mode=mode,
+            loading_mode=mode,
         )
         for inspection_mode in InspectionMode.values():
             file_name = f"test-{inspection_mode}.csv"